backend/requirements.txt
tenacity==8.2.2
flask==3.0.2
flask-cors==4.0.0
psycopg2==2.9.9
pandas==2.2.1
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
pydantic==2.3.0
python-dateutil==2.8.2
schedule==1.2.0
openpyxl==3.1.2
wheel==0.43.0